Add files via upload
Clean Code Use:
This code has descriptive names such as the functions sum, resta, multiplication, and division, it clearly describes the operation they perform.
It has small functions, which make the code easier to read and understand, it also has clear comments, does not contain unnecessary comments, and only includes them when necessary to clarify the purpose or operation of the code. It avoids code duplication, by using functions to perform common operations such as addition, subtraction, multiplication, and division. The error handling means that when the division is by zero it won´t show an error code and besides there is a proper use of whitespace it´s consistent.